Cast sizeof to int in formatted string precision arguments
Cast the result of the sizeof operator to int when using it as a
precision argument for a formatted string, in case size_t is not
the same as (unsigned) int.

+++ src/fileio.cpp
@@ -764,10 +764,10 @@ bool TarScanner::AddFile(const char *filename, size_t basepath_length, const cha
 		char name[sizeof(th.prefix) + 1 + sizeof(th.name) + 1];
 		if (th.prefix[0] != '\0') {
 			bstrfmt (name, "%.*s" PATHSEP "%.*s",
-				sizeof(th.prefix), th.prefix,
-				sizeof(th.name), th.name);
+				(int)sizeof(th.prefix), th.prefix,
+				(int)sizeof(th.name), th.name);
 		} else {
-			bstrfmt (name, "%.*s", sizeof(th.name), th.name);
+			bstrfmt (name, "%.*s", (int)sizeof(th.name), th.name);
 		}
 
 		/* Calculate the size of the file.. for some strange reason this is stored as a string */
